% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kmeans_HW2Version.R
\name{MyKmeans_R}
\alias{MyKmeans_R}
\title{R version.}
\usage{
MyKmeans_R(X, K, M = NULL, numIter = 100)
}
\arguments{
\item{X}{n by p matrix containing n data points to cluster}

\item{K}{integer specifying number of clusters}

\item{M}{(optional) K by p matrix of cluster centers}

\item{numIter}{number of maximal iterations for the algorithm, the default value is 100}
}
\value{
Y
}
\description{
R version.
}
